:truck: [#1546] -- Move FormLogic model into it's own module
