+++ src/openforms/forms/models/__init__.py
@@ -1,9 +1,10 @@
 from .category import Category
-from .form import Form, FormLogic
+from .form import Form
 from .form_definition import FormDefinition
 from .form_step import FormStep
 from .form_variable import FormVariable
 from .form_version import FormVersion
+from .logic import FormLogic
 from .pricing_logic import FormPriceLogic
 
 __all__ = [
